[#17] Fixes to the CSV importer
csv_to_table_def.py now tries to detect date columns and set them to the
date datatype in the SQL create statment. To make this work a fix to the
param handling in the Drakefile was needed.

An optional variable can be added to the drake command to specify the
date format to look for. The default pattern is "%d-%m-%Y %H:%M:%S UTC".

Fix table field names creation. CSV column headers starting with an
integer now are prepended a "_", and empty headers are given the name
"columnN" where N is the column count.