[#17] Bugfixes
Bugfixes to the Drakefile to work with the different security of
Postgres on reporting.test.akvo.org.
Note that the env variable PGPASSWORD needs the password of the DB_USER
to be able to run the psql commands.

Also a small bugfix to the column types for postgres, using bigint
rather than integer to be able to handle larger ints.