[#17] Fix sqlite import
For some reason the header row is imported as data. I think this is new
sqlite behaviour, I'm pretty sure it did not work this way previously.

Anyway, to fix this the python script that converts the delimiters from
comma to tab now skips the first line of the CSV.